+++ setup.cfg
@@ -6,4 +6,4 @@ max-complexity = 20
 [isort]
 line_length = 100
 known_first_party = views
-known_third_party = aiohttp,click,pandas,pytest,yaml
+known_third_party = aiohttp,click,pytest,yaml
